dry up code by creating a module to implment the common parts of a gen server based line member
