+++ lib/warehouse_simulator/picker.ex
@@ -4,9 +4,7 @@ defmodule WarehouseSimulator.Picker do
   can and pass it down the line.
   """
 
-  require WarehouseSimulator.LineMember
-  alias WarehouseSimulator.LineMember
-  use GenServer
+  use WarehouseSimulator.GenServerLineMember
 
   def start_link(parameters) do
     GenServer.start_link(__MODULE__, parameters)
@@ -21,22 +19,6 @@ defmodule WarehouseSimulator.Picker do
     {:ok, state}
   end
 
-  def process_pick_ticket(picker, receive_at, pick_ticket, current_contents \\ %{}) do
-    GenServer.call(picker, {:process_pick_ticket, receive_at, pick_ticket, current_contents})
-  end
-
-  def get_and_put_next_line_member(member, next_in_line, module) do
-    GenServer.call(member, {:get_and_put_next_line_member, next_in_line, module})
-  end
-
-  def elapsed_time(member) do
-    GenServer.call(member, {:elapsed_time})
-  end
-
-  def idle_time(member) do
-    GenServer.call(member, {:idle_time})
-  end
-
   def handle_call({:process_pick_ticket, receive_at, pick_ticket, current_contents}, _from, state) do
     {duration, contents} =
       pick_duration_and_contents(state[:parameters], pick_ticket, current_contents)
@@ -49,19 +31,6 @@ defmodule WarehouseSimulator.Picker do
       duration
     )
     |> LineMember.line_member_reply(state)
-  end
-
-  def handle_call({:get_and_put_next_line_member, next_in_line, module}, _from, state) do
-    LineMember.get_and_put_next_line_member(state[:line_member], next_in_line, module)
-    |> LineMember.line_member_reply(state)
-  end
-
-  def handle_call({:elapsed_time}, _from, state) do
-    {:reply, state[:line_member].now, state}
-  end
-
-  def handle_call({:idle_time}, _from, state) do
-    {:reply, state[:line_member].idle_duration, state}
   end
 
   defp pick_duration_and_contents(parameters, pick_ticket, current_contents) do
